BREAKING CHANGE to Constructor/::create && example
I made a breaking change to the constructor and factory creation
functions to allow for enabling/disabling of almost all features
via the constructor.

A discussion was brought up regarding the chaining on /r/PHP and
it was mentioned that for programmatic usefulness, it would be
better to have an array of flags to set during the construction,
instead of testing for certain variables and then triggering the
function call to the YTAudio object.

As such, I've made it so that instead of having two optional
parameters, the constructor and factory creation function have a
single optional parameter (after the still required one).  This
MUST be an array, but can be either associated, indexed, or a mix.

I've added usage of this new constructor to the example.php file
and modified the original pattern to omit the optional parameters.

Both uses in the example.php file will output the same XHTML.

Additionally, added the ability to turn on/off HTTPS.  (On by def.)